Fixup compiler warning in patches.fixes/vfs-fix-race-between-fcntl-and-file-f_flags-checks.patch.
suse-commit: d758cc3bd692a0c53c25eebd51e1932025e3ac91

+++ fs/aio.c
@@ -449,7 +449,7 @@ static struct kiocb *__aio_get_req(struct kioctx *ctx)
 	req->ki_flags = 0;
 	req->ki_users = 2;
 	req->ki_key = 0;
-	req->ki_ctx = (unsigned long)ctx;
+	req->ki_ctx = ctx;
 	req->ki_cancel = NULL;
 	req->ki_retry = NULL;
 	req->ki_dtor = NULL;
